prima/route-planner/route_planner/requirements.txt
pytz==2021.3
psycopg2-binary==2.8.3
pandas==1.3.5
numpy==1.21.5
ortools==9.1.9490
geopy==2.2.0

google-cloud-pubsub==2.0.0
google-cloud-storage==2.0.0
google-cloud-core==2.0.0


#added for consumer
memory-profiler==0.60.0
openpyxl==3.0.9

memoization==0.4.0
google-cloud-optimization==1.0.0
protobuf==3.19.0

scipy==1.7.3
shapely==1.8.4
here-location-services==0.4.0
werkzeug==1.0.1
redis==3.3.8